requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
motor>=2.5.1
pydantic>=1.9.0
python-dotenv>=0.19.0
httpx>=0.19.0
chromadb>=0.4.13
langchain>=0.0.267
pytest>=6.2.5
pytest-asyncio>=0.16.0
black>=21.9b0
isort>=5.9.3
flake8>=3.9.2